requirements.txt
cycler==0.11.0
fonttools==4.34.4
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.1
packaging==21.3
pdf2image==1.16.0
Pillow==9.2.0
pyparsing==3.0.9
PyPDF2==2.10.0
python-dateutil==2.8.2
regex==2022.7.25
scipy==1.9.0
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.64.0
